Added group sorting functionality to Dataset and Point.
Added ability to sort Dataset into groups of points
based on a group hash calculated from the IDs.

Also changed the Dataset spec to have the pending groups
for the rest of the tests related to comparing files
with two stage groups.

Still need to create the test data files and manually determine
results for the comparisons before creating the test cases.